refactor: 소셜로그인을 할 경우 socailId, type, (nickname)만 저장하도록 변경
소셜 로그인 할 경우 이메일 및 비밀번호를 저장하지 않습니다.

Related to: #27
