+++ backend/src/main/java/com/prograngers/backend/dto/response/auth/naver/NaverUserInfoResponse.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.dto.response.auth.naver;
 
-import com.prograngers.backend.entity.Member;
-import com.prograngers.backend.entity.MemberType;
+import com.prograngers.backend.entity.member.Member;
+import com.prograngers.backend.entity.member.MemberType;
 import lombok.Data;
 
 @Data
@@ -10,12 +10,9 @@ public class NaverUserInfoResponse {
 
     public Member toMember(){
         return Member.builder()
-                .name(response.getName())
-                .email(response.getEmail())
                 .socialId(Long.valueOf(response.getId().hashCode()))
                 .type(MemberType.NAVER)
                 .nickname(response.getNickname())
-                .phoneNumber(response.getMobile())
                 .build();
     }
 }
